# openapi.yaml
openapi: 3.1.0
info:
  title: Writers Platform API
  description: A modern writing platform API built with FastAPI
  version: 1.0.0
paths:
...
  /api/v1/collections/:
    post:
      tags:
        - collections
      summary: Create Collection
      description: |-
        Create a new collection (series/book).

        Rate limit: 5 collections per minute per IP
        Requires: Valid JWT token, active user
      operationId: create_collection_api_v1_collections__post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - collections
      summary: Get Collections
      description: |-
        Get collections with filtering and pagination.

        Rate limit: 30 requests per minute per IP
        Authentication: Optional
      operationId: get_collections_api_v1_collections__get
      security:
        - HTTPBearer: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/app__models__collection__CollectionType'
              - type: 'null'
            title: Type
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/app__models__collection__CollectionStatus'
              - type: 'null'
            title: Status
        - name: author_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Author Id
        - name: is_featured
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: 'null'
            title: Is Featured
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            title: Skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionResponse'
                title: Response Get Collections Api V1 Collections  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/collections/{collection_id}:
    get:
      tags:
        - collections
      summary: Get Collection
      description: |-
        Get a specific collection with its articles.

        Rate limit: 60 requests per minute per IP
        Authentication: Optional (required for draft collections)
      operationId: get_collection_api_v1_collections__collection_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            title: Collection Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionWithAuthor'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
        - collections
      summary: Update Collection
      description: |-
        Update an existing collection.

        Rate limit: 10 updates per minute per IP
        Requires: Valid JWT token, collection ownership or admin
      operationId: update_collection_api_v1_collections__collection_id__put
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            title: Collection Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - collections
      summary: Delete Collection
      description: |-
        Delete a collection.

        Rate limit: 5 deletions per minute per IP
        Requires: Valid JWT token, collection ownership or admin
      operationId: delete_collection_api_v1_collections__collection_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            title: Collection Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/collections/slug/{slug}:
    get:
      tags:
        - collections
      summary: Get Collection By Slug
      description: |-
        Get a specific collection by slug with its articles.

        Rate limit: 60 requests per minute per IP
        Authentication: Optional (required for draft collections)
      operationId: get_collection_by_slug_api_v1_collections_slug__slug__get
      security:
        - HTTPBearer: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionWithAuthor'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/collections/user/{user_id}:
    get:
      tags:
        - collections
      summary: Get User Collections
      description: |-
        Get collections by a specific user.

        Rate limit: 30 requests per minute per IP
        Authentication: Optional (required to see drafts)
      operationId: get_user_collections_api_v1_collections_user__user_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/app__models__collection__CollectionStatus'
              - type: 'null'
            title: Status
        - name: type
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/app__models__collection__CollectionType'
              - type: 'null'
            title: Type
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            title: Skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionResponse'
                title: Response Get User Collections Api V1 Collections User  User Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/collections/{collection_id}/publish:
    post:
      tags:
        - collections
      summary: Publish Collection
      description: |-
        Publish a draft collection.

        Rate limit: 10 publishes per minute per IP
        Requires: Valid JWT token, collection ownership or admin
      operationId: publish_collection_api_v1_collections__collection_id__publish_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            title: Collection Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CollectionCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
          description: Collection title
        description:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Description
          description: Collection description
        type:
          $ref: '#/components/schemas/app__schemas__collection__CollectionType'
          description: Collection type (series or book)
        cover_image:
          anyOf:
            - type: string
            - type: 'null'
          title: Cover Image
          description: Cover image URL
      type: object
      required:
        - title
        - type
      title: CollectionCreate
      description: Schema for creating a new collection.
    CollectionResponse:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
          description: Collection title
        description:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Description
          description: Collection description
        type:
          $ref: '#/components/schemas/CollectionType-Output'
          description: Collection type (series or book)
        cover_image:
          anyOf:
            - type: string
            - type: 'null'
          title: Cover Image
          description: Cover image URL
        id:
          type: integer
          title: Id
        slug:
          type: string
          title: Slug
        status:
          $ref: '#/components/schemas/CollectionStatus-Output'
        author_id:
          type: integer
          title: Author Id
        article_count:
          type: integer
          title: Article Count
          description: Number of articles in collection
          default: 0
        total_views:
          type: integer
          title: Total Views
          description: Total views across all articles
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
        published_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Published At
      type: object
      required:
        - title
        - type
        - id
        - slug
        - status
        - author_id
        - created_at
      title: CollectionResponse
      description: Schema for collection responses.
    CollectionStatus-Output:
      type: string
      enum:
        - draft
        - published
        - archived
      title: CollectionStatus
      description: Collection status enumeration.
...
    CollectionType-Output:
      type: string
      enum:
        - series
        - book
      title: CollectionType
      description: Collection type enumeration.
    CollectionUpdate:
      properties:
        title:
          anyOf:
            - type: string
              maxLength: 200
              minLength: 1
            - type: 'null'
          title: Title
        description:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Description
        type:
          anyOf:
            - $ref: '#/components/schemas/app__schemas__collection__CollectionType'
            - type: 'null'
        status:
          anyOf:
            - $ref: '#/components/schemas/app__schemas__collection__CollectionStatus'
            - type: 'null'
        cover_image:
          anyOf:
            - type: string
            - type: 'null'
          title: Cover Image
      type: object
      title: CollectionUpdate
      description: Schema for updating an existing collection.
    CollectionWithAuthor:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
          description: Collection title
        description:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Description
          description: Collection description
        type:
          $ref: '#/components/schemas/CollectionType-Output'
          description: Collection type (series or book)
        cover_image:
          anyOf:
            - type: string
            - type: 'null'
          title: Cover Image
          description: Cover image URL
        id:
          type: integer
          title: Id
        slug:
          type: string
          title: Slug
        status:
          $ref: '#/components/schemas/CollectionStatus-Output'
        author_id:
          type: integer
          title: Author Id
        article_count:
          type: integer
          title: Article Count
          description: Number of articles in collection
          default: 0
        total_views:
          type: integer
          title: Total Views
          description: Total views across all articles
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
        published_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Published At
        author:
          $ref: '#/components/schemas/app__schemas__user__UserResponse'
      type: object
      required:
        - title
        - type
        - id
        - slug
        - status
        - author_id
        - created_at
        - author
      title: CollectionWithAuthor
      description: Schema for collection with author information.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    app__models__collection__CollectionStatus:
      type: string
      enum:
        - draft
        - published
        - completed
        - archived
      title: CollectionStatus
      description: Collection publication status.
    app__models__collection__CollectionType:
      type: string
      enum:
        - series
        - book
        - anthology
        - course
      title: CollectionType
      description: Collection type enum.
...
    app__schemas__collection__CollectionStatus:
      type: string
      enum:
        - draft
        - published
        - archived
      title: CollectionStatus
      description: Collection status enumeration.
    app__schemas__collection__CollectionType:
      type: string
      enum:
        - series
        - book
      title: CollectionType
      description: Collection type enumeration.
    app__schemas__user__UserResponse:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        full_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Full Name
        avatar_url:
          anyOf:
            - type: string
            - type: 'null'
          title: Avatar Url
      type: object
      required:
        - id
        - username
      title: UserResponse
      description: Basic user response schema for article relations.
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
    HTTPBearer:
      type: http
      scheme: bearer